pyproject.toml
[project]
name = "homework-30"
version = "0.1.0"
description = ""
authors = [
    {name = "Egor",email = "epoplavskiy223@gmail.com"}
]
readme = "README.md"
requires-python = "^3.14"
dependencies = [
    "django (>=6.0.4,<7.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2-binary (>=2.9.12,<3.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "django-filter (>=25.2,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
lint = [
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]
